# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Time Tracking API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /resource-bookings/release-past-bulk:
    post:
      operationId: releasePastBulk
      tags: [resourceBookings]
      summary: Bulk-release past undelivered plan for unreleased bookings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReleasePastBulkBody"
      responses:
        "200":
          description: List of bookings that were released
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReleasePastBulkResult"
  /resource-bookings/{id}/release-past:
    post:
      operationId: releasePastBooking
      tags: [resourceBookings]
      summary: Release past undelivered plan for a single booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceBookingAction"
        "404":
          description: Booking not found
  /resource-bookings/{id}/unrelease:
    post:
      operationId: unreleaseBooking
      tags: [resourceBookings]
      summary: Undo release of past undelivered plan for a single booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceBookingAction"
        "404":
          description: Booking not found
...
components:
  schemas:
...
    ResourceBookingAction:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        projectId:
          type: integer
        projectRoleId:
          type: ["integer", "null"]
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        hoursPerDay:
          type: number
        pastReleasedAt:
          type: ["string", "null"]
          format: date-time
      required:
        - id
        - employeeId
        - projectId
        - startDate
        - endDate
        - hoursPerDay
    ReleasePastBulkBody:
      type: object
      properties:
        projectId:
          type: integer
        employeeId:
          type: integer
        dryRun:
          type: boolean
    ReleasePastBulkResult:
      type: object
      properties:
        released:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              employeeId:
                type: integer
              employeeName:
                type: ["string", "null"]
              projectId:
                type: integer
              projectName:
                type: string
              projectRoleId:
                type: ["integer", "null"]
              projectRoleName:
                type: ["string", "null"]
              startDate:
                type: string
                format: date
              endDate:
                type: string
                format: date
              pastReleasedAt:
                type: ["string", "null"]
                format: date-time
                nullable: true
              pastUndeliveredDays:
                type: number
            required:
              - id
              - employeeId
              - projectId
              - projectName
              - startDate
              - endDate
              - pastUndeliveredDays
      required:
        - released
...