# lib/api-zod/src/generated/types/resourceBookingAction.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Time Tracking API
 * OpenAPI spec version: 0.1.0
 */

export interface ResourceBookingAction {
  id: number;
  employeeId: number;
  projectId: number;
  /** @nullable */
  projectRoleId?: number | null;
  startDate: Date;
  endDate: Date;
  hoursPerDay: number;
  /** @nullable */
  pastReleasedAt?: Date | null;
}